Exercise 4.1
clean_session = true (for both publisher and subscriber)
qos = 0 (for both publisher and subscriber)
Write down your expected result. Then, follow the process above and note the actual results - are they as expected? Explain.

The results are equivalent to the expected, because the messages are not delivred to the csubscribers, since we specified that the message maybe will not be delivred (qos = 0)