Exercise 4.3
Write down your expected result.
Then, follow the process above and note the actual results - are they as expected? Explain.

The client was not removed from the server upon desconnection, so when connected again with a message with priority of at minimum one delivree, it was sent to the client