Exercise 4.5
clean_session = false for the subscriber and true for the publisher; qos = 1 for both clients ? The message was delivred, because the subscriber remain even upon desconnection, so when a message with right prority arrives It will be delivred.
same as above, except qos = 0 for the publisher ? No message was delivred, because none of them has the priority to be delivred.
same as above, with qos = 2 for both clients ? The message was delivred, since we stablished that the its priority was 2, by that it should be delivred exactly once time