Exercise 4.8
Find a way to test the functionning of the Last Will Testament (Cf. lecture notes or online documentation).
Hints: have a publisher connect to the broker setting a last will message.
Have the publisher send messages to a topic and at least one subscriber listening to that topic.
Interrupt the publisher's connection abruptly, without giving it a chance to send a disconnect message (e.g. stop its process: kill -9).
Check if the subscriber has received any message...

In this case using mqtt spy the las will message was received