Exercise 6.1
Find a way to send a CONNECT binary message to the JoramMQ Broker; and to display the returned message (i.e. CONNACK).
Do not use any of the support libraries (e.g. paho) and tools (e.g. mqtt-spy) employed in the previous exercises -- the objective here is to work directly with MQTT binary messages, as in the MQTT standard specification. You may use any binary editting tool and communication means (over TCP) to format/decipher and send/receive the MQTT messages to/from the Broker.